+++ crates/bevy_flair_css_parser/src/reflect.rs
@@ -6,10 +6,10 @@ mod image;
 mod text;
 mod ui;
 
-pub(crate) use color::parse_color;
+pub use color::parse_color;
 pub(crate) use enums::parse_enum_as_property_value;
 pub(crate) use grid::{parse_grid_track_vec, parse_repeated_grid_track_vec};
-pub(crate) use ui::parse_val;
+pub use ui::parse_val;
 
 use crate::error::CssError;
 use bevy_app::{App, Plugin};