+++ crates/bevy_flair_css_parser/src/reflect/enums.rs
@@ -3,7 +3,7 @@ use crate::error::CssError;
 use crate::error_codes::enums as error_codes;
 use crate::reflect::ReflectParseCssEnum;
 use crate::utils::parse_property_value_with;
-use bevy_flair_core::{PropertyValue, ReflectValue};
+use bevy_flair_core::PropertyValue;
 use bevy_reflect::{
     DynamicEnum, DynamicVariant, Enum, EnumInfo, FromReflect, FromType, TypeInfo, Typed,
     VariantInfo,
@@ -102,7 +102,7 @@ pub fn parse_enum_value<T: FromReflect + Typed + Enum>(parser: &mut Parser) -> R
 pub fn parse_enum_as_property_value<T: FromReflect + Typed + Enum>(
     parser: &mut Parser,
 ) -> Result<PropertyValue, CssError> {
-    parse_property_value_with(parser, parse_enum_value::<T>).map(|p| p.map(ReflectValue::new))
+    parse_property_value_with(parser, parse_enum_value::<T>).map(PropertyValue::into_reflect_value)
 }
 
 impl<T> FromType<T> for ReflectParseCssEnum